+++ fixtures/citigoe_iv_2020_ne14e2_style.yaml
@@ -134,6 +134,8 @@ reports:
           power: 61
           type: null
           capacity_in_liters: null
+        gearbox:
+          type: E1F  
         model: CitigoE iV
         title: Škoda CitigoE iV
         manufacturing_date: '2019-12-10'
