+++ fixtures/djapollo_enyaq_5AZJK2_85.yaml
@@ -241,6 +241,8 @@ reports:
         capacity_in_liters: null
         power: 210
         type: iV
+      gearbox:
+        type: E1H
       manufacturing_date: '2024-09-24'
       max_charging_power: 125
       model: Enyaq
