+++ fixtures/elroq_2025_PYLQK2_Sportline.yaml
@@ -242,6 +242,8 @@ reports:
         capacity_in_liters: null
         power: 210
         type: iV
+      gearbox:
+        type: E1H
       manufacturing_date: '2025-06-03'
       max_charging_power: 125
       model: Elroq
