+++ fixtures/enyaq_2023_5AZJJ2_80.yaml
@@ -205,6 +205,8 @@ reports:
           power: 150
           type: iV
           capacity_in_liters: null
+        gearbox:
+          type: E1H
         model: Enyaq
         title: Škoda Enyaq
         manufacturing_date: '2023-04-24'
