+++ fixtures/enyaq_2024_5AZJK2_85.yaml
@@ -233,6 +233,8 @@ reports:
         capacity_in_liters: null
         power: 210
         type: iV
+      gearbox:
+        type: E1H
       manufacturing_date: '2024-04-07'
       max_charging_power: 125
       model: Enyaq
