+++ fixtures/enyaq_2025_5azjk2_85.yaml
@@ -224,6 +224,8 @@ reports:
         capacity_in_liters: null
         power: 210
         type: iV
+      gearbox:
+        type: E1H
       manufacturing_date: '2025-03-31'
       max_charging_power: 125
       model: Enyaq
