+++ fixtures/kamiq_2021_NW44KD_crossover.yaml
@@ -137,6 +137,8 @@ reports:
         capacity_in_liters: 1.5
         power: 110
         type: TSI
+      gearbox:
+        type: A7F
       manufacturing_date: '2021-02-10'
       max_charging_power: null
       model: Kamiq
