+++ fixtures/karoq_2023_NU74MD_se_l.yaml
@@ -143,6 +143,8 @@ reports:
         capacity_in_liters: 1.5
         power: 110
         type: TSI
+      gearbox:
+        type: A7F
       manufacturing_date: '2023-10-06'
       max_charging_power: null
       model: Karoq
