+++ fixtures/octavia_2023_NX34JD_style.yaml
@@ -204,6 +204,8 @@ reports:
         capacity_in_liters: 1.5
         power: 110
         type: TSI
+      gearbox:
+        type: A7F
       manufacturing_date: '2023-01-18'
       max_charging_power: null
       model: Octavia
