+++ fixtures/sonar98_enyaq_2022_5AZJJ2_80.yaml
@@ -240,6 +240,8 @@ reports:
         capacity_in_liters: null
         power: 150
         type: iV
+      gearbox:
+        type: E1H
       manufacturing_date: '2022-03-24'
       max_charging_power: 125
       model: Enyaq
