+++ fixtures/superb_iv_2020_3v35xc_liftback_l_and_k.yaml
@@ -225,6 +225,8 @@ reports:
         capacity_in_liters: 1.4
         power: 160
         type: TSI iV
+      gearbox:
+        type: A6F
       manufacturing_date: '2020-06-08'
       max_charging_power: null
       model: Superb
