+++ fixtures/superb_iv_2021_3v54xc_style.yaml
@@ -234,6 +234,8 @@ reports:
         capacity_in_liters: 1.4
         power: 160
         type: TSI iV
+      gearbox:
+        type: A6F
       manufacturing_date: '2020-09-04'
       max_charging_power: null
       model: Superb
