+++ fixtures/superb_iv_2025_nz5ddd_combi_edition_130.yaml
@@ -203,6 +203,8 @@ reports:
         capacity_in_liters: 1.5
         power: 110
         type: TSI MHEV
+      gearbox:
+        type: A7F
       manufacturing_date: '2025-11-21'
       max_charging_power: null
       model: Superb
