+++ fixtures/webspider_enyaq_2021_5AZJJ2_80.yaml
@@ -230,6 +230,8 @@ reports:
         capacity_in_liters: null
         power: 150
         type: iV
+      gearbox:
+        type: E1H
       exterior_colour: null
       exterior_dimensions: null
       manufacturing_date: '2021-04-08'
